+++ example/url_example.py
@@ -1,5 +1,5 @@
 from funmovie.utils import *
 
-url = 'thunder://QUFtYWduZXQ6P3h0PXVybjpidGloOjY5NjVhMWE4MDRjYTY4MGNjMjRhNmU5OTEwNDNhMzY5YjFhMDViNzlaWg=='
+url = "thunder://QUFtYWduZXQ6P3h0PXVybjpidGloOjY5NjVhMWE4MDRjYTY4MGNjMjRhNmU5OTEwNDNhMzY5YjFhMDViNzlaWg=="
 print(url)
 print(thunder2url(url))