+++ funmovie/database/__init__.py
@@ -1,1 +1,0 @@
-from funmovie.database.core import MovieManage, MagnetManage