+++ src/funmovie/__init__.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 # @Time    : 2019/04/02 20:01
 # @Author  : niuliangtao
 # @Site    :