+++ src/funmovie/database/__init__.py
@@ -1,0 +1,1 @@
+from funmovie.database.core import MagnetManage, MovieManage